requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
pikepdf==8.14.0
httpx==0.25.2
Werkzeug==3.0.1
gunicorn==21.2.0
